requirements.txt
Flask
Jinja2
flask_wtf
flask-cors
gunicorn
openai
flask_seeder
flask_sqlalchemy
flask_migrate
redis
pymysql
python-dotenv
flask-restplus
Flask-OAuthlib
Flask-Login
Flask-WTF
Flask-Bootstrap
py_vapid
pywebpush
flask_crontab
apscheduler
unittest2
marshmallow
flask-admin
flask_paginate
apscheduler
discord.py
Flask-Uploads
Pillow
pytesseract
opencv-python
scikit-image
matplotlib
PyJWT
bcrypt
flask_bcrypt
flask_jwt_extended
setuptools
Werkzeug
requests